+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/core/ERC20.kt
@@ -2,10 +2,11 @@ package io.horizontalsystems.ethereumkit.core
 
 import io.horizontalsystems.ethereumkit.EthereumKit
 import io.horizontalsystems.ethereumkit.EthereumKit.KitState
+import java.math.BigDecimal
 
 class ERC20(var listener: EthereumKit.ListenerERC20) {
 
-    var balance: Double = 0.0
+    var balance: BigDecimal = BigDecimal.valueOf(0.0)
 
     var kitState: KitState = KitState.NotSynced
         set(value) {
